pyproject.toml
[tool.poetry]
name = "car-service-crm"
version = "0.1.0"
description = ""
authors = ["shulika-w <shulika.w@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.41"
aiosqlite = "^0.21.0"
pydantic = "^2.11.7"
python-multipart = "^0.0.20"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
jose = {extras = ["cryptography"], version = "^1.0.0"}
pymysql = "^1.1.1"
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
greenlet = "^3.2.3"
bcrypt = "<4.1"
python-dotenv = "^1.1.1"
aiosmtplib = "^4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.1.0"
fastapi = {extras = ["all"], version = "^0.116.1"}

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
